rename functions

+++ oneflow/core/actor/actor.h
@@ -115,7 +115,7 @@ class Actor {
   Regst* GetNaiveFirstCurReadable();
   Regst* GetSoleProducedRegst(int64_t regst_desc_id);
 
-  virtual int32_t WritingFreeProducedRegstDescNum() const { return 0; }
+  virtual int64_t WritingFreeProducedRegstDescNum() const { return 0; }
 
  private:
   bool IsReadReady();
